refactor: consolidate shared data structures
