+++ src/plcc/load_spec/load_rough_spec/__init__.py
@@ -1,5 +1,6 @@
-from .parse_lines import Line, parse_lines
-from .parse_blocks import Block, parse_blocks, UnclosedBlockError
-from .parse_includes import Include, parse_includes
-from .parse_dividers import Divider, parse_dividers
+from ..structs import Block, Divider, Include, Line
+from .parse_lines import parse_lines
+from .parse_blocks import parse_blocks, UnclosedBlockError
+from .parse_includes import parse_includes
+from .parse_dividers import parse_dividers
 from .parse_rough import parse_rough
