+++ src/plcc/load_spec/load_rough_spec/load_rough_spec_test.py
@@ -1,10 +1,10 @@
 from pytest import raises, mark, fixture
 
-from .parse_lines import Line
-from .parse_blocks import Block
-from .parse_dividers import Divider, parse_dividers
+from ..structs import Line, RoughSpec
+from ..structs import Block, Divider
+from .parse_dividers import parse_dividers
 from .load_rough_spec import load_rough_spec
-from .split_rough_spec import RoughSpec, split_rough_spec
+from .split_rough_spec import split_rough_spec
 
 
 def test_load_rough_spec(fs):
