+++ src/plcc/load_spec/load_rough_spec/parse_lines_test.py
@@ -1,7 +1,9 @@
 from pytest import raises, mark, fixture
 
+from ..structs import Line
 
-from .parse_lines import parse_lines, Line
+
+from .parse_lines import parse_lines
 
 
 def test_None_yields_nothing():
