+++ src/plcc/load_spec/load_rough_spec/split_rough_spec_test.py
@@ -1,9 +1,11 @@
 from pytest import raises, mark, fixture
 
-from .parse_lines import Line
-from .parse_dividers import Divider, parse_dividers
+from ..structs import Divider, RoughSpec
+
+from ..structs import Line
+from .parse_dividers import parse_dividers
 from .load_rough_spec import load_rough_spec
-from .split_rough_spec import RoughSpec, split_rough_spec
+from .split_rough_spec import split_rough_spec
 
 
 def test_empty():
