+++ src/plcc/load_spec/parse_spec/__init__.py
@@ -1,7 +1,8 @@
+from ..structs import CodeFragment, Line, SemanticSpec, TargetLocator
 from .parse_syntactic_spec import parse_syntactic_spec
 from .parse_lexical_spec import parse_lexical_spec
 
-from ..load_rough_spec.parse_lines import Line, parse_lines
+from ..load_rough_spec.parse_lines import parse_lines
 from ..load_rough_spec.parse_rough import parse_rough
 
-from .parse_semantic_spec import parse_semantic_spec, SemanticSpec, CodeFragment, TargetLocator
+from .parse_semantic_spec import parse_semantic_spec
