+++ src/plcc/load_spec/parse_spec/parse_semantic_spec/__init__.py
@@ -1,3 +1,4 @@
-from .parse_semantic_spec import SemanticSpec, parse_semantic_spec
-from .parse_code_fragments import CodeFragment, parse_code_fragments, UndefinedTargetLocatorError, DuplicateTargetLocatorError, CodeFragmentMissingBlockError
-from .parse_target_locator import TargetLocator, parse_target_locator
+from ...structs import CodeFragment, SemanticSpec, TargetLocator
+from .parse_semantic_spec import parse_semantic_spec
+from .parse_code_fragments import parse_code_fragments, UndefinedTargetLocatorError, DuplicateTargetLocatorError, CodeFragmentMissingBlockError
+from .parse_target_locator import parse_target_locator
