+++ src/plcc/load_spec/parse_spec/parse_semantic_spec/parse_code_fragments.py
@@ -1,14 +1,10 @@
-from dataclasses import dataclass
-from .parse_target_locator import TargetLocator, parse_target_locator
-from plcc.load_spec.load_rough_spec.parse_lines import Line
-from plcc.load_spec.load_rough_spec.parse_blocks import Block
-from plcc.load_spec.load_rough_spec.parse_dividers import Divider
-import re
 
-@dataclass
-class CodeFragment:
-    targetLocator: TargetLocator
-    block: Block
+from plcc.load_spec.structs import CodeFragment
+from .parse_target_locator import parse_target_locator
+from plcc.load_spec.structs import Line
+from plcc.load_spec.structs import Block
+from plcc.load_spec.structs import Divider
+import re
 
 def parse_code_fragments(lines_and_blocks: list[Line | Block]):
     parser = CodeFragmentParser(lines_and_blocks)
