+++ src/plcc/load_spec/parse_spec/parse_semantic_spec/parse_semantic_spec.py
@@ -1,15 +1,10 @@
-from dataclasses import dataclass
-from .parse_code_fragments import CodeFragment, parse_code_fragments
-from plcc.load_spec.load_rough_spec.parse_lines import Line
-from plcc.load_spec.load_rough_spec.parse_blocks import Block
-from plcc.load_spec.load_rough_spec.parse_dividers import Divider
-import re
 
-@dataclass
-class SemanticSpec:
-    language: str
-    tool: str
-    codeFragmentList: [CodeFragment]
+from plcc.load_spec.structs import SemanticSpec
+from .parse_code_fragments import parse_code_fragments
+from plcc.load_spec.structs import Line
+from plcc.load_spec.structs import Block
+from plcc.load_spec.structs import Divider
+import re
 
 def parse_semantic_spec(semantic_spec: list[Divider | Line | Block]) -> SemanticSpec:
     divider = semantic_spec[0]
