+++ src/plcc/load_spec/validate_spec/validate_lexical_spec/__init__.py
@@ -1,8 +1,9 @@
+from ...structs import Block, Divider, Include, LexicalRule, LexicalSpec, Line
 from .validate_lexical_spec import validate_lexical_spec
 
-from ...parse_spec.parse_lexical_spec import parse_lexical_spec, LexicalRule, LexicalSpec
-from ...load_rough_spec.parse_lines import Line, parse_lines
-from ...load_rough_spec.parse_blocks import Block, parse_blocks, UnclosedBlockError
-from ...load_rough_spec.parse_includes import Include, parse_includes
-from ...load_rough_spec.parse_dividers import Divider, parse_dividers
+from ...parse_spec.parse_lexical_spec import parse_lexical_spec
+from ...load_rough_spec.parse_lines import parse_lines
+from ...load_rough_spec.parse_blocks import parse_blocks, UnclosedBlockError
+from ...load_rough_spec.parse_includes import parse_includes
+from ...load_rough_spec.parse_dividers import parse_dividers
 from ...load_rough_spec.parse_rough import parse_rough
