+++ src/plcc/load_spec/validate_spec/validate_lexical_spec/validate_lexical_spec_test.py
@@ -1,9 +1,11 @@
 from pytest import raises, mark, fixture
 
-from ...load_rough_spec.parse_lines import Line
-from .validate_lexical_spec import ValidationError, validate_lexical_spec
-from ...parse_spec.parse_lexical_spec import LexicalRule, LexicalSpec
-from .lexical_errors import ValidationError, InvalidNameFormatError, DuplicateNameError, InvalidPatternError, InvalidRuleError
+from ...errors import DuplicateNameError, InvalidNameFormatError, InvalidPatternError, ValidationError
+
+from ...structs import LexicalRule, Line
+from .validate_lexical_spec import validate_lexical_spec
+from ...structs import LexicalSpec
+from ...errors import InvalidRuleError
 
 def test_empty_no_errors():
     lexicalSpec = makeLexicalSpec([])
