+++ src/plcc/load_spec/validate_spec/validate_semantic_spec/__init__.py
@@ -1,3 +1,4 @@
+from ...structs import Line
 from .validate_semantic_spec import validate_semantic_spec
 
-from ...load_rough_spec.parse_lines import Line, parse_lines
+from ...load_rough_spec.parse_lines import parse_lines
