+++ src/plcc/load_spec/validate_spec/validate_syntactic_spec/check_ll1/build_spec_grammar_test.py
@@ -1,10 +1,11 @@
 from pytest import raises
+from ....errors import InvalidSyntacticSpecException
 from plcc.load_spec.load_rough_spec.parse_dividers import parse_dividers
-from plcc.load_spec.load_rough_spec.parse_lines import Line
+from plcc.load_spec.structs import Line
 from .LL1Wrapper import wrap_ll1
 from .build_spec_grammar import build_spec_grammar
 from plcc.load_spec.parse_spec.parse_syntactic_spec.parse_syntactic_spec import parse_syntactic_spec
-from .errors import InvalidSymbolException, InvalidSyntacticSpecException
+from ....errors import InvalidSymbolException
 
 
 def test_init():
