+++ src/plcc/load_spec/validate_spec/validate_syntactic_spec/check_ll1/check_parsing_table_for_ll1.py
@@ -1,14 +1,7 @@
 from collections import defaultdict
-from dataclasses import dataclass
+
+from plcc.load_spec.errors import LL1Error
 from .build_parsing_table import Table
-
-
-@dataclass
-class LL1Error:
-    def __init__(self, cell, production):
-        self.cell = cell
-        self.production = production
-        self.message = f"Two production rules in the same parsing table cell: {cell} -> {production}"
 
 
 def check_parsing_table_for_ll1(parsingTable: Table) -> list[LL1Error]:
